WIP: make Node-version to test memory usage
