+++ .gitignore
@@ -1,4 +1,3 @@
-# Dependencies
 node_modules/
-
+dist
 .env
