+++ src/utils/createCommand.ts
@@ -1,4 +1,4 @@
-import type { Command } from "@/types";
 import type { ChatInputCommandInteraction } from "discord.js";
+import type { Command } from "#types.js";
 
 export default <T = ChatInputCommandInteraction>(props: Command<T>): Command<T> => props;
