+++ src/utils/getCommandUsed.ts
